# use more efficient wire serialization