--- a/app/app_test.go
+++ b/app/app_test.go
@@ -162,6 +162,7 @@ func newStartedApp(
 			},
 			BlockOnSend:            true,
 			DisableGzipCompression: true,
+			EnableMsgpackEncoding:  true,
 			Metrics:                sdPeer,
 		},
 	})