Switch to msgpack for internal (peer to peer and upstream) communication.

Switch to jsoniter for batch event unmarshal, which is notably faster than stock json.

The perf win here is pretty clear, and this benchmark doesn't capture the gains from encoding upstream traffic with msgpack, since libhoney's Writer transmission always uses json.

```
benchmark                              old ns/op     new ns/op     delta
BenchmarkDistributedTraces/batch-8     25991         17567         -32.41%

benchmark                              old allocs     new allocs     delta
BenchmarkDistributedTraces/batch-8     296            226            -23.65%

benchmark                              old bytes     new bytes     delta
BenchmarkDistributedTraces/batch-8     23634         19938         -15.64%
```